Update2

+++ TP3-Layers-ej2/src/ar/unrn/tp3/persistence/RegistroDeEmpleadosEnDisco.java
@@ -14,8 +14,10 @@ import ar.unrn.tp3.modelo.RegistroDeEmpleados;
 
 public class RegistroDeEmpleadosEnDisco implements RegistroDeEmpleados {
 
+	@Override
 	public void registrar(Empleado emp) {
 		try {
+			if(!existeUnEmpleado(emp))
 			Files.write(Paths.get("C:\\Users\\ELJUEEN\\Desktop\\archivoTP3empleados.txt"), emp.toString().getBytes(),
 					StandardOpenOption.APPEND);
 		} catch (IOException e) {
@@ -23,6 +25,7 @@ public class RegistroDeEmpleadosEnDisco implements RegistroDeEmpleados {
 		}
 	}
 
+	@Override
 	public ArrayList<Empleado> obtenerEmpleados() {
 		try {
 			List<String> lineaCadaEmpleado = Files.readAllLines(Paths.get("C:\\Users\\ELJUEEN\\Desktop\\archivoTP3empleados.txt"));
@@ -47,4 +50,33 @@ public class RegistroDeEmpleadosEnDisco implements RegistroDeEmpleados {
 		return null;
 	}
 
+	@Override
+	public boolean existeUnEmpleado(Empleado emp) {
+		try {
+			List<String> lineaCadaEmpleado = Files.readAllLines(Paths.get("C:\\Users\\ELJUEEN\\Desktop\\archivoTP3empleados.txt"));
+			ArrayList<Empleado> empleados = new ArrayList<Empleado>();
+			for (String s : lineaCadaEmpleado) {
+				String[] split = s.split(",");
+				String fecha = split[2];
+				String[] split2 = fecha.split("-");
+				LocalDate cumple = LocalDate.of(Integer.parseInt(split2[0]), Integer.parseInt(split2[1]),
+						Integer.parseInt(split2[2]));
+				Empleado empl = new Empleado(split[0], split[1], cumple, split[3]);
+				empleados.add(empl);
+
+			}
+			
+			for (Empleado e : empleados) {
+				if(e.obtenerEmail().equals(emp.obtenerEmail())) {
+					return true;
+				}
+			}
+			
+			
+		} catch (IOException e) {
+			System.out.println("No se pueden obtener los empleados desde el archivo.");
+		}
+
+		return false;
+	}
 }
